Address feedback from code review
