+++ app/models/reports.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "reports/monthly_bulk_serializer"
+
+module Reports; end
