+++ app/services/reports/monthly_bulk_report_service.rb
@@ -17,6 +17,9 @@ module Reports
         # retry 3 times then exit with failure
       # end
       # rubocop:enable Layout/CommentIndentation
+    rescue StandardError => e
+      Airbrake.notify e, file_name: file_name
+      Rails.logger.error "Generate bulk export csv error for #{file_name}:\n#{error}"
     ensure
       temp_file.close
       temp_file.unlink
@@ -29,8 +32,8 @@ module Reports
     end
 
     def file_name
-      data_from_date = @first_day_of_the_month.strftime("%Y%m%d")
-      data_to_date = @first_day_of_the_month.end_of_month.strftime("%Y%m%d")
+      data_from_date = @first_day_of_the_month.to_formatted_s(:plain_year_month_day)
+      data_to_date = @first_day_of_the_month.end_of_month.to_formatted_s(:plain_year_month_day)
 
       "#{data_from_date}-#{data_to_date}.csv"
     end
